refactor: MyChallenge 클래스에 Facade 패턴 적용
- MyChallengeService를 Facade 클래스로 변경
- Facade 인터페이스 선언 및 구현
