+++ src/main/java/com/genius/gitget/challenge/myChallenge/controller/MyChallengeController.java
@@ -7,7 +7,7 @@ import com.genius.gitget.challenge.myChallenge.dto.ActivatedResponse;
 import com.genius.gitget.challenge.myChallenge.dto.DoneResponse;
 import com.genius.gitget.challenge.myChallenge.dto.PreActivityResponse;
 import com.genius.gitget.challenge.myChallenge.dto.RewardRequest;
-import com.genius.gitget.challenge.myChallenge.service.MyChallengeService;
+import com.genius.gitget.challenge.myChallenge.facade.MyChallengeFacadeService;
 import com.genius.gitget.global.security.domain.UserPrincipal;
 import com.genius.gitget.global.util.response.dto.ListResponse;
 import com.genius.gitget.global.util.response.dto.SingleResponse;
@@ -28,14 +28,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @CrossOrigin
 public class MyChallengeController {
-    private final MyChallengeService myChallengeService;
+    private final MyChallengeFacadeService myChallengeFacadeService;
 
 
     @GetMapping("/my/pre-activity")
     public ResponseEntity<ListResponse<PreActivityResponse>> getPreActivityChallenges(
             @AuthenticationPrincipal UserPrincipal userPrincipal
     ) {
-        List<PreActivityResponse> preActivityInstances = myChallengeService.getPreActivityInstances(
+        List<PreActivityResponse> preActivityInstances = myChallengeFacadeService.getPreActivityInstances(
                 userPrincipal.getUser(),
                 DateUtil.convertToKST(LocalDateTime.now()));
 
@@ -49,7 +49,7 @@ public class MyChallengeController {
     public ResponseEntity<ListResponse<ActivatedResponse>> getActivatedChallenges(
             @AuthenticationPrincipal UserPrincipal userPrincipal
     ) {
-        List<ActivatedResponse> activatedInstances = myChallengeService.getActivatedInstances(
+        List<ActivatedResponse> activatedInstances = myChallengeFacadeService.getActivatedInstances(
                 userPrincipal.getUser(),
                 DateUtil.convertToKST(LocalDateTime.now()));
 
@@ -62,7 +62,7 @@ public class MyChallengeController {
     public ResponseEntity<ListResponse<DoneResponse>> getDoneChallenges(
             @AuthenticationPrincipal UserPrincipal userPrincipal
     ) {
-        List<DoneResponse> doneInstances = myChallengeService.getDoneInstances(
+        List<DoneResponse> doneInstances = myChallengeFacadeService.getDoneInstances(
                 userPrincipal.getUser(),
                 DateUtil.convertToKST(LocalDateTime.now()));
 
@@ -78,7 +78,7 @@ public class MyChallengeController {
     ) {
         LocalDate kstDate = DateUtil.convertToKST(LocalDateTime.now());
         RewardRequest rewardRequest = new RewardRequest(userPrincipal.getUser(), instanceId, kstDate);
-        DoneResponse doneResponse = myChallengeService.getRewards(rewardRequest, false);
+        DoneResponse doneResponse = myChallengeFacadeService.getRewards(rewardRequest, false);
 
         return ResponseEntity.ok().body(
                 new SingleResponse<>(SUCCESS.getStatus(), SUCCESS.getMessage(), doneResponse)
