+++ src/main/java/com/genius/gitget/challenge/myChallenge/facade/MyChallengeFacade.java
@@ -1,0 +1,16 @@
+package com.genius.gitget.challenge.myChallenge.facade;
+
+import com.genius.gitget.challenge.myChallenge.dto.ActivatedResponse;
+import com.genius.gitget.challenge.myChallenge.dto.DoneResponse;
+import com.genius.gitget.challenge.myChallenge.dto.PreActivityResponse;
+import com.genius.gitget.challenge.user.domain.User;
+import java.time.LocalDate;
+import java.util.List;
+
+public interface MyChallengeFacade {
+    List<PreActivityResponse> getPreActivityInstances(User user, LocalDate targetDate);
+
+    List<ActivatedResponse> getActivatedInstances(User user, LocalDate targetDate);
+
+    List<DoneResponse> getDoneInstances(User user, LocalDate targetDate);
+}
